peerdas: Split `helpers.go` in multiple files respecting the specification.
