--- beacon-chain/core/peerdas/log.go
@@ -1,5 +1,0 @@
-package peerdas
-
-import "github.com/sirupsen/logrus"
-
-var log = logrus.WithField("prefix", "peerdas")
